rfe-gui/requirements.txt
pyserial
PyQt6

